backend/requirements.txt
fastapi
uvicorn
requests
openai
python-dotenv
openai>=1.0.0


